[Bugfix 20514] Extensions: Add a way to launch sample stacks
Sample stacks can be launched from the Extension Manager. Launched stacks
have their filename set to empty to the user is propted to choose a
location to save to is they save they stack. This ensures the original
sample stack is always opened from the Extension Manager and the user can
make changes and save the stacks.

It is assumed that the sample stacks will be in a "samples" folder in the
extension folder.